move files, game switching inital data
