+++ src/marioparty/utils/Gamepad.java
@@ -1,4 +1,4 @@
-package utils;
+package marioparty.utils;
 
 import net.java.games.input.Component;
 import net.java.games.input.Component.Identifier;
